refactor(multipooler): move replication methods into it's own file, adds unit tests
